+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move "getContentString" method to Localization service [#2291](https://github.com/rokwire/illinois-app/issues/2291).
 ### Added
 - Cache Canvas Courses [#2294](https://github.com/rokwire/illinois-app/issues/2294).
+- Added courses caching in Student Courses service [#2303](https://github.com/rokwire/illinois-app/issues/2303).
 
 ## [4.1.21] - 2022-08-17
 ### Changed